feat(canvas): retire the home-canvas concept
The channel landing page is a native view already, and the only navigation
into a home canvas had no callers, so the auto-created seeded home board was
vestigial: UI shipped as per-channel data that can't be updated fleet-wide,
makes channel-open depend on the build pipeline, and hides query failures as
"No canvases yet." (how the missing system.canvases table read as data loss).

Remove the client machinery (ensure/reset/seed, tRPC routes, isHome plumbing),
the backend is_home field, filter, and unique constraint (constraint dropped
for real; column removed from Django state only with a db-side default first),
and soft-delete the system-generated home boards. User-created canvases are
untouched. Generated OpenAPI/MCP types and snapshots regenerated.

Generated-By: PostHog Code
Task-Id: d75a71da-b662-4101-af83-0588664849c2
