+++ products/desktop/packages/host-router/src/routers/dashboards.router.ts
@@ -10,7 +10,6 @@ import {
   createDashboardInput,
   dashboardIdInput,
   dashboardRecordSchema,
-  ensureHomeCanvasInput,
   listDashboardsInput,
   renameDashboardInput,
   revertCanvasInput,
@@ -113,22 +112,6 @@ export const dashboardsRouter = router({
     .output(dashboardRecordSchema)
     .mutation(({ ctx, input }) =>
       ctx.container.get<IDashboardsService>(DASHBOARDS_SERVICE).rename(input),
-    ),
-  ensureHomeCanvas: publicProcedure
-    .input(ensureHomeCanvasInput)
-    .output(dashboardRecordSchema)
-    .mutation(({ ctx, input }) =>
-      ctx.container
-        .get<IDashboardsService>(DASHBOARDS_SERVICE)
-        .ensureHomeCanvas(input.channelId),
-    ),
-  resetHomeCanvas: publicProcedure
-    .input(ensureHomeCanvasInput)
-    .output(dashboardRecordSchema)
-    .mutation(({ ctx, input }) =>
-      ctx.container
-        .get<IDashboardsService>(DASHBOARDS_SERVICE)
-        .resetHomeCanvas(input.channelId),
     ),
   delete: publicProcedure
     .input(dashboardIdInput)
