+++ services/mcp/src/api/generated.ts
@@ -13289,7 +13289,6 @@ export namespace Schemas {
       readonly pinned: boolean;
       /** @nullable */
       readonly pinned_at: string | null;
-      readonly is_home: boolean;
       /**
          * Id of the live source version — pass as expected_current_version_id on publish. Null before the first publish.
          * @nullable
@@ -13494,8 +13493,6 @@ export namespace Schemas {
          * @maxLength 64
          */
       template_id?: string;
-      /** Create the canvas as the channel's home board (at most one per channel). */
-      is_home?: boolean;
     }
 
     /**
@@ -77660,10 +77657,6 @@ export namespace Schemas {
      * Only return canvases in this channel.
      */
     channel?: string;
-    /**
-     * Filter by channel-home status.
-     */
-    is_home?: boolean;
     /**
      * Number of results to return per page.
      */
